+++ examples/platform/nrf528xx/app/include/Server.h
@@ -28,5 +28,7 @@ using DemoSessionManager = chip::SecureSessionMgr<chip::Transport::UDP>;
 
 void StartServer(DemoSessionManager * sessions);
 void InitDataModelHandler();
+void SetDeviceName(const char * newDeviceName);
+void SendUDPBroadCast();
 
 #endif // NRF5_COMMON_SERVER_H
